refactor(winner): winner 서비스 없애고 domain 영역에서 구하도록 수정
