+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,6 @@
 package racingcar;
 
 import racingcar.controller.RacingCarController;
-import racingcar.service.WinnerService;
 import racingcar.validator.CarNameValidator;
 import racingcar.validator.TryCountValidator;
 import racingcar.view.InputView;
@@ -10,12 +9,9 @@ import racingcar.view.OutputView;
 public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView(new TryCountValidator(), new CarNameValidator());
-        WinnerService winnerService = new WinnerService();
         OutputView outputView = new OutputView();
 
-        RacingCarController racingCarController = new RacingCarController(
-                inputView, winnerService, outputView
-        );
+        RacingCarController racingCarController = new RacingCarController(inputView, outputView);
 
         racingCarController.playGame();
     }
